Move all configs to a central directory
