+++ packages/validator/vitest.spec.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../vitest.base.unit.config";
+import vitestConfig from "../../configs/vitest.base.unit.config";
 
 export default mergeConfig(
   vitestConfig,
